factor inner classes out of ThriftStructMetadataBuilder
