+++ swift-codec/src/main/java/com/facebook/swift/codec/metadata/FieldExtractor.java
@@ -1,0 +1,59 @@
+/*
+ * Copyright (C) 2012 Facebook, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License. You may obtain
+ * a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
+package com.facebook.swift.codec.metadata;
+
+import com.facebook.swift.codec.ThriftField;
+
+import java.lang.reflect.Field;
+import java.lang.reflect.Type;
+
+class FieldExtractor extends Extractor
+{
+    private final Field field;
+
+    FieldExtractor(Field field, ThriftField annotation)
+    {
+        super(annotation);
+        this.field = field;
+    }
+
+    public Field getField()
+    {
+        return field;
+    }
+
+    @Override
+    public String extractName()
+    {
+        return field.getName();
+    }
+
+    @Override
+    public Type getJavaType()
+    {
+        return field.getGenericType();
+    }
+
+    @Override
+    public String toString()
+    {
+        final StringBuilder sb = new StringBuilder();
+        sb.append("FieldExtractor");
+        sb.append("{field=").append(field);
+        sb.append('}');
+        return sb.toString();
+    }
+}
